+++ bocadillo/app_types.py
@@ -1,4 +1,4 @@
-from typing import Any, Awaitable, Callable, MutableMapping, Union, TypeVar
+from typing import Any, Awaitable, Callable, MutableMapping, TypeVar
 
 from .request import Request
 from .response import Response
@@ -23,9 +23,7 @@ class ASGIApp:
 
 
 # HTTP
-AsyncHandler = Callable[[Request, Response, Any], Awaitable[None]]
-SyncHandler = Callable[[Request, Response, Any], None]
-Handler = Union[AsyncHandler, SyncHandler]
+Handler = Callable[[Request, Response, Any], Awaitable[None]]
 ErrorHandler = Callable[[Request, Response, _E], Awaitable[None]]