+++ bocadillo/compat.py
@@ -1,15 +1,6 @@
-import asyncio
+import inspect
 import re
-from typing import (
-    Callable,
-    cast,
-    List,
-    TypeVar,
-    Union,
-    Optional,
-    Any,
-    Awaitable,
-)
+from typing import Callable, List, TypeVar, Any
 
 try:
     # >= 3.7
@@ -22,14 +13,23 @@ except ImportError:  # pragma: no cover
         yield enter_result
 
 
-from starlette.concurrency import run_in_threadpool
-
 _CAMEL_REGEX = re.compile(r"(.)([A-Z][a-z]+)")
 _SNAKE_REGEX = re.compile(r"([a-z0-9])([A-Z])")
 
 _V = TypeVar("_V")
 
 
+class ExpectedAsync(Exception):
+    """Raised when a feature expected an asynchronous function."""
+
+
+def check_async(func, reason: str):
+    if not inspect.iscoroutinefunction(func):
+        message = f"{reason}. Hint: use `async def` instead of `def`."
+        raise ExpectedAsync(message)
+
+
+# For 3.6 compatibility (only available in 3.7+).
 class asyncnullcontext:
     def __init__(self, enter_result: Any = None):
         self._result = enter_result
@@ -39,32 +39,6 @@ class asyncnullcontext:
 
     async def __aexit__(self, *args):
         pass
-
-
-async def call_async(
-    func: Union[Callable[..., _V], Callable[..., Awaitable[_V]]],
-    *args: Any,
-    sync: Optional[bool] = None,
-    **kwargs: Any
-) -> _V:
-    """Call a function in an async manner.
-
-    # Parameters
-    func:
-        a callable that is either awaited (if a coroutine function)
-        or run in the thread pool (if a regular function).
-    sync (bool):
-        a hint as to whether `func` is synchronous. If not given, it is
-        inferred as `asyncio.iscoroutinefunction(func)`.
-
-    # See Also
-    - [Executing code in thread or process pools](https://docs.python.org/3/library/asyncio-eventloop.html#executing-code-in-thread-or-process-pools)
-    """
-    if sync or (sync is None and not asyncio.iscoroutinefunction(func)):
-        return await run_in_threadpool(func, *args, **kwargs)
-
-    async_func = cast(Callable[..., Awaitable[_V]], func)
-    return await async_func(*args, **kwargs)
 
 
 def camel_to_snake(name: str) -> str: