+++ bocadillo/hooks.py
@@ -1,8 +1,8 @@
 import inspect
 from functools import wraps
-from typing import Callable, cast, Dict, Union, Awaitable, Type
+from typing import Callable, Dict, Union, Awaitable, Type
 
-from .compat import call_async
+from .compat import check_async
 from .request import Request
 from .response import Response
 from .routing import HTTPRoute
@@ -36,29 +36,43 @@ class Hooks:
         """
         return self._hook_decorator(AFTER, hook, *args, **kwargs)
 
+    @staticmethod
+    def _prepare(hook, *args, **kwargs) -> HookFunction:
+        class_based = not inspect.isfunction(hook)
+
+        if class_based:
+            assert hasattr(
+                hook, "__call__"
+            ), "class-based hooks must implement __call__()"
+
+        check_target = hook.__call__ if class_based else hook
+        name = (
+            hook.__class__.__name__ + ".__call__"
+            if class_based
+            else hook.__name__
+        )
+        check_async(check_target, reason=f"hook '{name}' must be asynchronous")
+
+        # Enclose args and kwargs
+        async def hook_func(req: Request, res: Response, params: dict):
+            await hook(req, res, params, *args, **kwargs)
+
+        return hook_func
+
     def _hook_decorator(
         self, hook_type: str, hook: HookFunction, *args, **kwargs
     ):
-        # Enclose args and kwargs
-        async def hook_func(req: Request, res: Response, params: dict):
-            await call_async(  # type: ignore
-                hook, req, res, params, *args, **kwargs
-            )
+        hook = self._prepare(hook, *args, **kwargs)
 
-        def decorator(handler: Union[Type[View], Handler]):
-            """Attach the hook to the given handler."""
-            if not inspect.isclass(handler):
-                return _with_hook(hook_type, hook_func, cast(Handler, handler))
+        def attach_hook(handler: Union[Type[View], Handler]):
+            if inspect.isclass(handler):
+                # Recursively apply the hook to all handlers.
+                for method, _handler in get_handlers(handler).items():
+                    setattr(handler, method, attach_hook(_handler))
+                return handler
+            return _with_hook(hook_type, hook, handler)
 
-            view_cls = cast(View, handler)
-
-            # Recursively apply hook to all handlers.
-            for method, _handler in get_handlers(view_cls).items():
-                setattr(view_cls, method, decorator(_handler))
-
-            return view_cls
-
-        return decorator
+        return attach_hook
 
 
 def _with_hook(hook_type: str, func: HookFunction, handler: Handler):
@@ -70,14 +84,14 @@ def _with_hook(hook_type: str, func: HookFunction, handler: Handler):
             req, res = args[1:3]
         assert isinstance(req, Request)
         assert isinstance(res, Response)
-        await call_async(func, req, res, kw)
+        await func(req, res, kw)
 
     if hook_type == BEFORE:
 
         @wraps(handler)
         async def with_before_hook(*args, **kwargs):
             await call_hook(args, kwargs)
-            await call_async(handler, *args, **kwargs)
+            await handler(*args, **kwargs)
 
         return with_before_hook
 
@@ -85,13 +99,13 @@ def _with_hook(hook_type: str, func: HookFunction, handler: Handler):
 
     @wraps(handler)
     async def with_after_hook(*args, **kwargs):
-        await call_async(handler, *args, **kwargs)
+        await handler(*args, **kwargs)
         await call_hook(args, kwargs)
 
     return with_after_hook
 
 
 # Pre-bind to module
-hooks = Hooks()
-before = hooks.before
-after = hooks.after
+_HOOKS = Hooks()
+before = _HOOKS.before
+after = _HOOKS.after