+++ tests/test_http_views.py
@@ -1,7 +1,39 @@
 import pytest
 
-from bocadillo import App, view
+from bocadillo import App, ExpectedAsync, view
 from bocadillo.constants import ALL_HTTP_METHODS
+
+
+def test_function_based_view(app: App, client):
+    @app.route("/")
+    @view()
+    async def index(req, res):
+        pass
+
+    assert client.get("/").status_code == 200
+
+
+def test_async_check(app):
+    def index(req, res):
+        pass
+
+    with pytest.raises(ExpectedAsync):
+        app.route("/")(index)
+
+    class Index:
+        def get(self, req, res):
+            pass
+
+    with pytest.raises(ExpectedAsync):
+        app.route("/")(Index)
+
+
+def test_view_decorator_is_optional(app: App, client):
+    @app.route("/")
+    async def index(req, res):
+        pass
+
+    assert client.get("/").status_code == 200
 
 
 def test_can_register_class_based_view(app: App):
@@ -43,32 +75,6 @@ def test_if_handle_is_implemented_then_bypasses_other_methods(app: App, client):
     response = client.get("/")
     assert response.status_code == 200
     assert response.text == "Handle!"
-
-
-def test_sync_handler(app: App, client):
-    @app.route("/")
-    class Index:
-        def get(self, req, res):
-            pass
-
-    assert client.get("/").status_code == 200
-
-
-def test_function_based(app: App, client):
-    @app.route("/")
-    @view()
-    def index(req, res):
-        pass
-
-    assert client.get("/").status_code == 200
-
-
-def test_view_decorator_is_optional(app: App, client):
-    @app.route("/")
-    def index(req, res):
-        pass
-
-    assert client.get("/").status_code == 200
 
 
 def test_from_obj(app: App, client):