+++ tests/utils.py
@@ -7,38 +7,18 @@ import requests
 
 
 @contextmanager
-def function_hooks(before_value: Any = True, after_value: Any = True):
+def function_hooks(expected_before: Any = True, expected_after: Any = True):
     flags = {"before": False, "after": False}
 
-    def before(req, res, params, value=before_value):
+    async def before(req, res, params, value=expected_before):
         nonlocal flags
         assert not flags["after"]
         flags["before"] = value
 
-    def after(req, res, params, value=after_value):
+    async def after(req, res, params, value=expected_after):
         nonlocal flags
         assert flags["before"]
         flags["after"] = value
-
-    yield before, after
-
-    assert flags["before"] == before_value
-    assert flags["after"] == after_value
-
-
-@contextmanager
-def async_function_hooks(expected_before=True, expected_after=True):
-    flags = {"before": False, "after": False}
-
-    async def before(req, res, params):
-        nonlocal flags
-        assert not flags["after"]
-        flags["before"] = True
-
-    async def after(req, res, params):
-        nonlocal flags
-        assert flags["before"]
-        flags["after"] = True
 
     yield before, after
 
@@ -55,7 +35,7 @@ def class_hooks():
             self.flag = flag
             self.value = value
 
-        def __call__(self, req, res, params):
+        async def __call__(self, req, res, params):
             nonlocal flags
             flags[self.flag] = self.value